requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.32.3
python-jose==3.3.0
python-multipart==0.0.6
pydantic==2.4.2
gunicorn==21.2.0
sqlalchemy==2.0.28
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9 